fix: rename tr app to config.
- moved trade_remedies_api -> config
- updated all related refs
- tidied README
