--- .all-contributorsrc
@@ -1,100 +1,0 @@
-{
-  "projectName": "trade-remedies-api",
-  "projectOwner": "uktrade",
-  "repoType": "github",
-  "repoHost": "https://github.com",
-  "files": [
-    "README.md"
-  ],
-  "imageSize": 100,
-  "commit": true,
-  "commitConvention": "none",
-  "contributors": [
-    {
-      "login": "harel",
-      "name": "Harel Malka",
-      "avatar_url": "https://avatars3.githubusercontent.com/u/985978?v=4",
-      "profile": "http://www.harelmalka.com/",
-      "contributions": [
-        "code",
-        "doc",
-        "review"
-      ]
-    },
-    {
-      "login": "bobmeredith",
-      "name": "Robert Meredith",
-      "avatar_url": "https://avatars2.githubusercontent.com/u/11422209?v=4",
-      "profile": "https://github.com/bobmeredith",
-      "contributions": [
-        "code",
-        "design"
-      ]
-    },
-    {
-      "login": "Luisella21",
-      "name": "Luisella Strona",
-      "avatar_url": "https://avatars1.githubusercontent.com/u/36708790?v=4",
-      "profile": "https://github.com/Luisella21",
-      "contributions": [
-        "review"
-      ]
-    },
-    {
-      "login": "markhigham",
-      "name": "Mark Higham",
-      "avatar_url": "https://avatars1.githubusercontent.com/u/2064710?v=4",
-      "profile": "https://github.com/markhigham",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "nao360",
-      "name": "Nao Yoshino",
-      "avatar_url": "https://avatars3.githubusercontent.com/u/6898065?v=4",
-      "profile": "https://github.com/nao360",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "ulcooney",
-      "name": "Paul Cooney",
-      "avatar_url": "https://avatars0.githubusercontent.com/u/1695475?v=4",
-      "profile": "https://github.com/ulcooney",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "michalc",
-      "name": "Michal Charemza",
-      "avatar_url": "https://avatars1.githubusercontent.com/u/13877?v=4",
-      "profile": "http://charemza.name/",
-      "contributions": [
-        "code",
-        "review"
-      ]
-    },
-    {
-      "login": "krishnawhite",
-      "name": "Krishna White",
-      "avatar_url": "https://avatars1.githubusercontent.com/u/5566533?v=4",
-      "profile": "https://github.com/krishnawhite",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "bquinn",
-      "name": "Brendan Quinn",
-      "avatar_url": "https://avatars0.githubusercontent.com/u/309976?v=4",
-      "profile": "http://blog.clueful.com.au/",
-      "contributions": [
-        "code"
-      ]
-    }
-  ],
-  "contributorsPerLine": 7
-}
