+++ .circleci/config.yml
@@ -33,7 +33,7 @@ jobs:
       - image: cimg/python:3.9.5
         environment:
           DATABASE_URL: psql://ubuntu@localhost:5432/trade_remedies
-          DJANGO_SETTINGS_MODULE: trade_remedies_api.settings.test
+          DJANGO_SETTINGS_MODULE: config.settings.test
           DJANGO_SECRET_KEY: ci_only
           ALLOWED_HOSTS: "*"
       - image: circleci/postgres:9.6.5-alpine-ram
