+++ pii-secret-exclude.txt
@@ -1,3 +1,8 @@
+.pii-secret-hook/file_content/pii-secret-log
+.pii-secret-hook/ner/pii-secret-log
+pii_secret_exclude.txt
+pii_ner_exclude.txt
+ner_output_file.txt
 requirements/base.txt
 runtime.txt
 setup/init.sql
