+++ trade_remedies_api/config/.coveragerc
