+++ trade_remedies_api/config/wsgi.py
@@ -11,7 +11,7 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "trade_remedies_api.settings.local")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings.local")
 
 # from whitenoise.django import DjangoWhiteNoise  # noqa
