+++ trade_remedies_api/core/services/auth.py
@@ -23,7 +23,7 @@ from security.constants import (
     ENVIRONMENT_GROUPS,
 )
 from core.constants import CONTENT_EMAIL_EXISTS
-from trade_remedies_api.version import __version__
+from config.version import __version__
 from .base import TradeRemediesApiView, ResponseSuccess, ResponseError
 from .exceptions import InvalidRequestParams, AccessDenied, InvalidRequestLockout
