--- trade_remedies_api/trade_remedies_api/constants.py
@@ -1,4 +1,0 @@
-STATE_UNKONWN = "UNKNOWN"
-STATE_INCOMPLETE = "INCOMPLETE"
-STATE_INPROGRESS = "INPROGRESS"
-STATE_COMPLETE = "COMPLETE"
